requirements.txt
# Core Dependencies
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.3.0

# Agent SDK
# Note: Replace with actual package name if using a specific agents framework
# agents-sdk>=1.0.0

# Type Checking
typing-extensions>=4.8.0

# Development Dependencies
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1
black>=23.11.0
mypy>=1.7.0
flake8>=6.1.0

# Optional: For enhanced API handling
httpx>=0.25.2
aiohttp>=3.9.1

# Optional: For CLI enhancements
rich>=13.7.0
click>=8.1.7
